Adding playwright tests
