+++ packages/scroll-spy/example/src/js/index.js
@@ -2,5 +2,5 @@ import scrollSpy from '../../../src';
     
 window.addEventListener('DOMContentLoaded', () => {
     scrollSpy('.js-scroll-spy', {single: true});
-    //scrollSpy('.js-scroll-spy', {single: false});
+    scrollSpy('.js-scroll-spy-multiple', {single: false});
 });
