+++ packages/scroll-spy/example/src/multiple.html
@@ -1,8 +1,8 @@
 <!doctype html>
 <html lang="en">
 	<head>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>StormID</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <style>
         * {
             margin:0;
@@ -82,13 +82,13 @@
 <body>
 	<header>
         <nav>
-            <a class="js-scroll-spy" href="#section1">Section 1</a>
-            <a class="js-scroll-spy" href="#section4">Section 4</a>
-            <a class="js-scroll-spy" href="#section5">Section 5</a>
-            <a class="js-scroll-spy" href="#section2">Section 2</a>
-            <a class="js-scroll-spy" href="#section3">Section 3</a>
-            <a class="js-scroll-spy" href="no-hash">No hash</a>
-            <a class="js-scroll-spy" href="#no-node">No node</a>
+            <a class="js-scroll-spy-multiple" href="#section1">Section 1</a>
+            <a class="js-scroll-spy-multiple" href="#section4">Section 4</a>
+            <a class="js-scroll-spy-multiple" href="#section5">Section 5</a>
+            <a class="js-scroll-spy-multiple" href="#section2">Section 2</a>
+            <a class="js-scroll-spy-multiple" href="#section3">Section 3</a>
+            <a class="js-scroll-spy-multiple" href="no-hash">No hash</a>
+            <a class="js-scroll-spy-multiple" href="#no-node">No node</a>
         </nav>
     </header>
     <main class="container">
