+++ packages/scroll-spy/tools/webpack.config.js
@@ -22,6 +22,11 @@ module.exports = {
             title: pkg.name,
             template: './example/src/index.html',
             filename: 'index.html'
+        }),
+        new HtmlWebpackPlugin({
+            title: pkg.name,
+            template: './example/src/multiple.html',
+            filename: 'multiple.html'
         })
     ],
     module: {
